+++ confik/src/third_party.rs
@@ -218,3 +218,49 @@ mod bigdecimal {
         type Builder = Option<Self>;
     }
 }
+
+#[cfg(feature = "ahash")]
+mod ahash {
+    use std::{fmt::Display, hash::Hash};
+
+    use ahash::{AHashMap, AHashSet};
+    use serde::de::DeserializeOwned;
+
+    use crate::{
+        helpers::{BuilderOf, KeyedContainer, KeyedContainerBuilder, UnkeyedContainerBuilder},
+        Configuration,
+    };
+
+    impl<T> Configuration for AHashSet<T>
+    where
+        T: Configuration + Hash + Eq,
+        BuilderOf<T>: Hash + Eq + 'static,
+    {
+        type Builder = UnkeyedContainerBuilder<AHashSet<BuilderOf<T>>, Self>;
+    }
+
+    impl<K, V> KeyedContainer for AHashMap<K, V>
+    where
+        K: Hash + Eq,
+    {
+        type Key = K;
+        type Value = V;
+
+        fn insert(&mut self, k: Self::Key, v: Self::Value) {
+            self.insert(k, v);
+        }
+
+        fn remove(&mut self, k: &Self::Key) -> Option<Self::Value> {
+            self.remove(k)
+        }
+    }
+
+    impl<K, V> Configuration for AHashMap<K, V>
+    where
+        K: Hash + Eq + Display + DeserializeOwned + 'static,
+        V: Configuration,
+        BuilderOf<V>: 'static,
+    {
+        type Builder = KeyedContainerBuilder<AHashMap<K, BuilderOf<V>>, Self>;
+    }
+}